// src/interfaces/user.ts
// Generated by https://quicktype.io

export interface LoginResponse {
  usuario: Usuario;
  token: string;
}

export interface RegisterResponse {
  usuario: Usuario;
  token: string;
}

export interface Usuario {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  uid: string;
}

export interface Usuario {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  uid: string;
}